pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-tag"
version = "0.1.0"
description = "Natural-language querying over IFC models using hybrid retrieval"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ifcopenshell>=0.8.4.post1",
    "logfire>=3.8.1",
    "nest-asyncio>=1.6.0",
    "networkx>=3.6.1",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "pydantic>=2.7.0",
    "pydantic-ai>=0.0.49",
    "python-dotenv>=1.2.1",
]

[project.scripts]
rag-tag = "rag_tag.run_agent:main"
rag-tag-ifc-to-csv = "rag_tag.parser.ifc_to_csv:main"
rag-tag-csv-to-sql = "rag_tag.parser.csv_to_sql:main"
rag-tag-csv-to-graph = "rag_tag.parser.csv_to_graph:main"

[dependency-groups]
dev = [
    "pre-commit>=3.7",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
line-ending = "lf"
